requirements_chatgpt.txt
google-api-python-client==2.108.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
openai==1.6.1
groq==0.4.1
python-dotenv==1.0.0
schedule==1.2.0
moviepy==1.0.3
Pillow==10.1.0
requests==2.31.0
